+++ test/test-load.js
@@ -2,7 +2,7 @@
 'use strict';
 var assert = require('assert');
 
-describe('qvext generator', function () {
+describe('qsExtension generator', function () {
   it('can be imported without blowing up', function () {
     var app = require('../app');
     assert(app !== undefined);