+++ ChangeLog.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## v0.0.3
+* Added Namespace 
+
+## v0.0.2 (2014-04-14)
+* Remove vsRegions completely
+* Added zip-support to build the final solution
+
 ## v0.0.1 (2014-03-21)
 * Added Changelog
 * Made vsRegions dynamic
+* Changes to _extensions.js (Script template)
+    * Added /*global define */
+    * Added 'use strict'
+    * Added jslint tolerance: **devel: true** (will be removed in grunt-build anyway)
+    * Added jslint tolerance: **white: true** = messy whitespace